backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==12.0

# Celery and Redis for async task processing
celery==5.3.4
redis==5.0.1

# Audio processing
openai-whisper
librosa
soundfile
pyannote.audio
silero-vad  # Voice Activity Detection

# ML and transformers
torch>=2.0.0
transformers
datasets

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0  # For async SQLite support

# Utilities
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
python-dotenv==1.0.0

# Optional: GPU acceleration
# torch-audio  # Uncomment if you want torchaudio features